Makefile: add ability to exclude features via environment.
The Makefile was overriding CPPFLAGS, which prohibited the
ability to change what features are compiled in by modifying
the environment.  This makes it usable in an environment
(such as an embedded device) where library support isn't
available.

Now, it will react to NO_SSL, NO_IPV6, and NO_LUAC. If any of
those are equal to "y", that feature will not be compiled in.